+++ provisioner/ansible-local/provisioner.go
@@ -232,6 +232,12 @@ func (p *Provisioner) executeAnsible(ui packer.Ui, comm packer.Communicator) err
 		return err
 	}
 	if cmd.ExitStatus != 0 {
+		if cmd.ExitStatus == 127 {
+			return fmt.Errorf("%s could not be found. Verify that it is available on the\n" +
+				"PATH after connecting to the machine.",
+				p.config.Command)
+		}
+
 		return fmt.Errorf("Non-zero exit status: %d", cmd.ExitStatus)
 	}
 	return nil